% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_ggdaisy_states.R
\name{get_ggdaisy_states}
\alias{get_ggdaisy_states}
\title{Get the states a phylogeny edge can be in.
These states are \code{\link{get_daisy_input_statuses}}
with an \code{invisible} state added}
\usage{
get_ggdaisy_states()
}
\description{
Get the states a phylogeny edge can be in.
These states are \code{\link{get_daisy_input_statuses}}
with an \code{invisible} state added
}
\author{
Richel J.C. Bilderbeek
}
