% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggd_create_phylo.R
\name{ggd_create_phylo}
\alias{ggd_create_phylo}
\title{Create a phylogeny with the correct statuses}
\usage{
ggd_create_phylo(clade_name, status, branching_times,
  island_age = max(branching_times))
}
\arguments{
\item{clade_name}{the clade's name}

\item{status}{the clade's status,
must be a member of \code{\link{get_daisy_input_statuses}}}

\item{branching_times}{branching times}

\item{island_age}{island age}
}
\value{
a phylogeny of class \code{phylo} with a \code{status} attribute
}
\description{
Create a phylogeny with the correct statuses
}
\author{
Richel J.C. Bilderbeek
}
