% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggd_create_phylo.R
\name{ggd_create_phylo_non_endemic_max_age}
\alias{ggd_create_phylo_non_endemic_max_age}
\title{Create a phylogeny with the correct statuses}
\usage{
ggd_create_phylo_non_endemic_max_age(time, taxon_label)
}
\value{
a phylogeny of class \code{phylo} with a \code{status} attribute
}
\description{
Create a phylogeny with the correct statuses
}
