% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggd_create_phylos.R
\name{ggd_create_phylos}
\alias{ggd_create_phylos}
\title{Creku
Plot a DAISIE data table}
\usage{
ggd_create_phylos(df, island_age)
}
\arguments{
\item{df}{a data frame}

\item{island_age}{age of the island}
}
\description{
Creku
Plot a DAISIE data table
}
\author{
Richel J.C. Bilderbeek
}
