% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggdaisy.R
\name{ggdaisie}
\alias{ggdaisie}
\title{Plot a DAISIE data table}
\usage{
ggdaisie(df, island_age)
}
\arguments{
\item{df}{a data frame}

\item{island_age}{age of the island}
}
\description{
Plot a DAISIE data table
}
\author{
Richel J.C. Bilderbeek
}
