% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_phylo_with_stem.R
\name{plot_phylo_with_stem_ggtree}
\alias{plot_phylo_with_stem_ggtree}
\title{Plot a phylogeny with a stem using ape}
\usage{
plot_phylo_with_stem_ggtree(phylo)
}
\arguments{
\item{phylo}{a phylogeny of class \code{phylo}}
}
\description{
Plot a phylogeny with a stem using ape
}
\author{
Richel J.C. Bilderbeek
}
