% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggd_brts_to_rnd_phylo.R
\name{ggd_brts_to_rnd_phylo}
\alias{ggd_brts_to_rnd_phylo}
\title{Create a comb phylo from branch length}
\usage{
ggd_brts_to_rnd_phylo(branching_times)
}
\description{
Create a comb phylo from branch length
}
