% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggd_create_phylo.R
\name{ggd_create_phylo}
\alias{ggd_create_phylo}
\title{Create a phylogeny with the correct statuses}
\usage{
ggd_create_phylo(clade_name, status, branching_times)
}
\description{
Create a phylogeny with the correct statuses
}
